1. fix word emphasis logic which capitalizes words that have a high sound intensity 2. fix suggested punctuation feature that infers punctuation based on pause lengths 3. made text method modifiers properties of Expression class 4. Made sound intensity methods use Double

+++ diction-processor/SoundIntensityDatum.swift
@@ -10,7 +10,7 @@ import Foundation
 
 struct SoundIntensityDatum {
     var date: Date
-    var intensity: Float
+    var intensity: Double
 
     var description: String {
         return "SoundIntensityDatum {\n\tdate: \(date)\n\tintensity:\(intensity)\n}"
@@ -20,7 +20,7 @@ struct SoundIntensityDatum {
        date = value
    }
    
-   mutating func setIntensity(value: Float) {
+   mutating func setIntensity(value: Double) {
        intensity = value
    }
 }
